requirements.txt

Flask==2.3.2
gunicorn==20.1.0
slackclient==2.9.4

